+++ beyond/CHANGELOG.md
@@ -1,11 +1,14 @@
+## 0.0.3
+- Fix override field name
+- Fix wrong fromJson binding on model sample 
+- Fix inserted records result
+- Sunseting DB insert method & fix create
+
 ## 0.0.2 
 
 - Fix env path
 - Change DB insert, to DB create
 - Update DB for multiple insert records
-- Fix override field name
-- Fix wrong fromJson binding on model sample 
-- Fix inserted records result
 
 ## 0.0.1